# swagger.yaml
openapi: 3.0.0
info:
  title: Muscle Feed
  version: 1.0.0
  description: API Documentation for Muscle Feed backend
components:
  schemas:
    UserDto:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        isVerified:
          type: boolean
        roles:
          type: array
          items:
            enum:
              - MODERATOR
              - ADMIN
        firstName:
          type: string
          example: Rob
        lastName:
          type: string
          example: Top
        phone:
          type: string
          example: +79123456789
        allergies:
          type: string
          example: "Морковь"
...
    DishTypeResponseDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: object
          properties:
            ru:
              type: string
              example: "Завтрак"
            he:
              type: string
              example: "ארוחת בוקר"
    DishResponseDto:
      type: object
      properties:
        id:
          type: integer
        adminName:
          type: string
          example: "Стейк с картофелем"
        dishType:
          $ref: "#/components/schemas/DishTypeResponseDto"
        picture:
          type: string
          example: "http://localhost:8080/static/7fbe6f80-800e-4520-9580-e60e3c93fc20.png"
        calories:
          type: integer
        weight:
          type: integer
        proteins:
          type: integer
        fats:
          type: integer
        carbohydrates:
          type: integer
        price:
          type: integer
        isPublished:
          type: boolean
        isIndividualOrderAvailable:
          type: boolean
        name:
          type: object
          properties:
            ru:
              type: string
              example: "Стейк с картофелем"
            he:
              type: string
              example: "סטייק עם תפוחי אדמה"
        description:
          type: object
          properties:
            ru:
              type: string
              example: "Сочный стейк с картофельным пюре и овощами."
            he:
              type: string
              example: "סטייק עסיסי עם פירה תפוחי אדמה וירקות."
        benefit:
          type: object
          properties:
            ru:
              type: string
              example: "Польза"
            he:
              type: string
              example: "יתרונות"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MenuResponseShortDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: object
          properties:
            ru:
              type: string
              example: "Сбалансированное меню"
            he:
              type: string
              example: "תפריט מאוזן"
        description:
          type: object
          properties:
            ru:
              type: string
              example: "Сбалансированное меню"
            he:
              type: string
              example: "תפריט מאוזן"
        calories:
          type: integer
        menuType:
          type: object
          properties:
            id:
              type: number
            name:
              type: object
              properties:
                name:
                  type: object
                  properties:
                    ru:
                      type: string
                    he:
                      type: string
            backgroundPicture:
              type: string
    PaymentMethodResponseDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: object
          properties:
            ru:
              type: string
              example: "Картой курьеру, при получении заказа"
            he:
              type: string
              example: "בכרטיס אשראי לשליח, ברגע קבלת המשלוח"
    OrderRequestDto:
      properties:
        menuId:
          type: integer
        startDate:
          type: string
          format: date
          example: 2025-02-17
        daysCount:
          type: integer
        skippedWeekdays:
          description: Numbers from 1 to 7 (1 - Mon, 2 Tues, etc.). It may be empty.
          type: array
          items:
            type: integer
        paymentMethodId:
          type: integer
        fullName:
          type: string
          example: "Иван Иванов"
        email:
          type: string
          format: email
        phone:
          type: string
          example: "+79876543210"
        allergies:
          type: string
          example: "Орехи"
        cityId:
          type: integer
        street:
          type: string
          example: "Ленина"
        house:
          type: string
          example: "10"
        floor:
          type: integer
        apartment:
          type: integer
        comment:
          type: string
          example: "Позвоните за 10 минут до доставки"
        promocodeId:
          type: integer
      required:
        - menuId
        - startDate
        - daysCount
        - skippedWeekdays
        - paymentMethodId
        - fullName
        - email
        - phone
        - cityId
        - street
        - house
    OrderRequestAdminDto:
      allOf:
        - $ref: "#/components/schemas/OrderRequestDto"
        - properties:
            userId:
              type: integer
            price:
              type: integer
            paidAmount:
              type: integer
            promocodeDiscount:
              type: integer
            giftDaysCount:
              type: integer
            finalPrice:
              type: integer
            freezeStartDate:
              type: string
              format: date
              example: 2025-02-17
            freezeEndDate:
              type: string
              format: date
              example: 2025-02-17
            isProcessed:
              type: boolean
            isAllowedExtendion:
              type: boolean
            isPaid:
              type: boolean
            isCompleted:
              type: boolean
          required:
            - price
            - finalPrice
            - isProcessed
            - isAllowedExtendion
            - isPaid
    OrderResponseDto:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        fullName:
          type: string
          example: Иван Ромашов
        email:
          type: string
          format: email
        phone:
          type: string
          example: +79123456789
        allergies:
          type: string
          example: Орехи
        finalPrice:
          type: integer
        menu:
          $ref: "#/components/schemas/MenuResponseShortDto"
        city:
          $ref: "#/components/schemas/CityResponseDto"
        street:
          type: string
          example: "Ленина"
        house:
          type: string
          example: "10"
        floor:
          type: integer
        apartment:
          type: integer
        comment:
          type: string
          example: "Позвоните за 10 минут до доставки"
        daysCount:
          type: integer
        giftDaysCount:
          type: integer
        daysLeft:
          type: integer
        startDate:
          type: string
          format: date
          example: 2025-02-17
        endDate:
          type: string
          format: date
          example: 2025-02-17
        skippedWeekdays:
          type: array
          items:
            type: integer
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethodResponseDto"
        isPaid:
          type: boolean
        isIndividual:
          type: boolean
        isFrozen:
          type: boolean
        freezeStartDate:
          type: string
          format: date
          example: 2025-02-17
        freezeEndDate:
          type: string
          format: date
          example: 2025-02-17
    TotalDto:
      type: object
      properties:
        calories:
          type: integer
        proteins:
          type: integer
        fats:
          type: integer
        carbohydrates:
          type: integer
    OrderChangeRequestResponseDto:
      type: object
      properties:
        id:
          type: integer
        orderChangeType:
          type: string
          enum:
            - MENU
            - CALORIES
            - DURATION
            - FORMAT
            - FREEZE
            - OTHER
        comment:
          type: string
          example: Заморозьте мой заказ
        isProcessed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updateddAt:
          type: string
          format: date-time
        order:
          $ref: "#/components/schemas/OrderRequestDto"
...
    CityResponseDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: object
          properties:
            ru:
              type: string
              example: "Хайфа"
            he:
              type: string
              example: "חיפה"
...
    PromocodeResponseDto:
      type: object
      properties:
        id:
          type: number
        code:
          type: string
          example: MUSCLE
        discount:
          type: integer
    SocialRequestDto:
      type: object
      properties:
        name:
          type: string
        link:
          type: string
        icon:
          type: string
          example: "http://localhost:8080/api/static/42c193f6-1ccc-4376-aef7-4f302817adc3.svg"
    SettingsResponseDto:
      type: object
      properties:
        id:
          type: number
        cycleStartDate:
          type: string
          format: date
          example: 2025-02-17
        phoneNumber:
          type: string
        email:
          type: string
        socials:
          type: array
          items:
            allOf:
              - type: object
                properties:
                  id:
                    type: number
              - $ref: "#/components/schemas/SocialRequestDto"
    PaginationBodyFields:
      type: object
      properties:
        totalCount:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
        isLast:
          type: boolean
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 5
    Status:
      name: status
      in: query
      required: false
      description: all - все заказы (по умолчанию), active - активные, individual - индивидуальные, frozen - замороженные, unpaid - неоплаченные, completed - завершенные, pending - ожидающие, terminating - завершающиеся, unprocessed - необработанные
      schema:
        type: string
        default: all
        enum:
          - all
          - active
          - individual
          - frozen
          - unpaid
          - completed
          - pending
          - terminating
          - unprocessed
  requestBodies:
...
    AdminOrder:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OrderRequestAdminDto"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []
paths:
...
  /api/admin/order:
    get:
      tags:
        - Admin Order
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Status"
        - name: search
          in: query
          required: false
          description: Поиск осуществляется по id, email, номер телефона, ФИО
          schema:
            type: string
            example: Иван
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                allOf:
                  - properties:
                      orders:
                        type: array
                        items:
                          $ref: "#/components/schemas/OrderResponseDto"
                  - $ref: "#/components/schemas/PaginationBodyFields"
    post:
      tags:
        - Admin Order
      requestBody:
        $ref: "#/components/requestBodies/AdminOrder"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  order:
                    $ref: "#/components/schemas/OrderResponseDto"
  /api/admin/order/{orderId}:
    get:
      tags:
        - Admin Order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  order:
                    type: object
                    properties:
                      id:
                        type: integer
                      fullName:
                        type: string
                        example: "Иван Иванов"
                      email:
                        type: string
                        format: email
                      phone:
                        type: string
                        example: "+79876543210"
                      allergies:
                        type: string
                        example: "Орехи"
                      street:
                        type: string
                        example: "Ленина"
                      house:
                        type: string
                        example: "10"
                      floor:
                        type: integer
                      apartment:
                        type: integer
                      comment:
                        type: string
                        example: "Позвоните за 10 минут до доставки"
                      price:
                        type: integer
                      paidAmount:
                        type: integer
                      promocodeDiscount:
                        type: integer
                      menuDiscount:
                        type: integer
                      finalPrice:
                        type: integer
                      isProcessed:
                        type: boolean
                      isAllowedExtendion:
                        type: boolean
                      isPaid:
                        type: boolean
                      isCompleted:
                        type: boolean
                      isIndividual:
                        type: boolean
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                      promocode:
                        $ref: "#/components/schemas/PromocodeResponseDto"
                      user:
                        $ref: "#/components/schemas/UserDto"
                      city:
                        $ref: "#/components/schemas/CityResponseDto"
                      paymentMethod:
                        $ref: "#/components/schemas/PaymentMethodResponseDto"
                      menu:
                        $ref: "#/components/schemas/MenuResponseShortDto"
                      daysCount:
                        type: integer
                      giftDaysCount:
                        type: integer
                      daysLeft:
                        type: integer
                      startDate:
                        type: string
                        format: date
                        example: 2025-02-17
                      endDate:
                        type: string
                        format: date
                        example: 2025-02-17
                      skippedWeekdays:
                        type: array
                        items:
                          type: integer
    patch:
      tags:
        - Admin Order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        $ref: "#/components/requestBodies/AdminOrder"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  order:
                    $ref: "#/components/schemas/OrderResponseDto"
    delete:
      tags:
        - Admin Order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  order:
                    $ref: "#/components/schemas/OrderResponseDto"
  /api/admin/order/{orderId}/day:
    get:
      tags:
        - Admin Order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  days:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        date:
                          type: string
                          format: date-time
                        isSkipped:
                          type: boolean
                        daySkipType:
                          type: string
                          enum:
                            - WEEKDAY_SKIPPED
                            - FROZEN
  /api/admin/order/day/{dayId}:
    get:
      tags:
        - Admin Order
      parameters:
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  dishes:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/DishResponseDto"
                        - properties:
                            count:
                              type: integer
                  total:
                    $ref: "#/components/schemas/TotalDto"
  /api/admin/order/day/{dayId}/replacement:
    get:
      tags:
        - Admin Order
      parameters:
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
        - name: dish_type_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  dishes:
                    type: array
                    items:
                      $ref: "#/components/schemas/DishResponseDto"
  /api/admin/order/select:
    post:
      tags:
        - Admin Order
      requestBody:
        content:
          application/json:
            schema:
              properties:
                dayId:
                  type: integer
                dishTypeId:
                  type: integer
                dishId:
                  type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  dish:
                    $ref: "#/components/schemas/DishResponseDto"
  /api/admin/order/change-request:
    get:
      tags:
        - Admin Order
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                allOf:
                  - properties:
                      orderChangeRequests:
                        type: array
                        items:
                          $ref: "#/components/schemas/OrderChangeRequestResponseDto"
                  - $ref: "#/components/schemas/PaginationBodyFields"
...
  /api/admin/order/change-request/{changeRequestId}:
    get:
      tags:
        - Admin Order
      parameters:
        - name: changeRequestId
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  orderChangeRequest:
                    $ref: "#/components/schemas/OrderChangeRequestResponseDto"
    patch:
      tags:
        - Admin Order
      parameters:
        - name: changeRequestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                isProcessed:
                  type: boolean
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  orderChangeRequest:
                    $ref: "#/components/schemas/OrderChangeRequestResponseDto"
...
  /api/order/{orderId}:
    get:
      tags:
        - Order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  order:
                    allOf:
                      - $ref: "#/components/schemas/OrderResponseDto"
                      - properties:
                          price:
                            type: integer
                          paidAmount:
                            type: integer
                          promocodeDiscount:
                            type: integer
                          menuDiscount:
                            type: integer
  /api/order/{orderId}/day:
    get:
      tags:
        - Order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  days:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        date:
                          type: string
                          format: date-time
                        isSkipped:
                          type: boolean
                        daySkipType:
                          type: string
                          enum:
                            - WEEKDAY_SKIPPED
                            - FROZEN
  /api/order/day/{dayId}:
    get:
      tags:
        - Order
      parameters:
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  dishes:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/DishResponseDto"
                        - properties:
                            count:
                              type: integer
                  total:
                    $ref: "#/components/schemas/TotalDto"
  /api/order/day/{dayId}/replacement:
    get:
      tags:
        - Order
      parameters:
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
        - name: dish_type_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  dishes:
                    type: array
                    items:
                      $ref: "#/components/schemas/DishResponseDto"
  /api/order/select:
    post:
      tags:
        - Order
      requestBody:
        content:
          application/json:
            schema:
              properties:
                dayId:
                  type: integer
                dishTypeId:
                  type: integer
                dishId:
                  type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  dish:
                    $ref: "#/components/schemas/DishResponseDto"
  /api/order/{orderId}/change-request:
    post:
      tags:
        - Order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                orderChangeType:
                  type: string
                  enum:
                    - MENU
                    - CALORIES
                    - DURATION
                    - FORMAT
                    - FREEZE
                    - OTHER
                comment:
                  type: string
                  example: Заморозьте мой заказ
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  orderChangeRequest:
                    $ref: "#/components/schemas/OrderChangeRequestResponseDto"
...
  /admin/settings/cycle-start-date:
    patch:
      tags:
        - Admin Settings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cycleStartDate:
                  type: string
                  format: date
                  example: 2025-02-17
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  settings:
                    $ref: "#/components/schemas/SettingsResponseDto"
  /admin/settings/contact:
    patch:
      tags:
        - Admin Settings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phoneNumber:
                  type: string
                email:
                  type: string
                socials:
                  type: array
                  items:
                    $ref: "#/components/schemas/SocialRequestDto"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  settings:
                    $ref: "#/components/schemas/SettingsResponseDto"
...